Change syntax to require less parentheses in thm/defthm hypotheses and conclusions
For thm and defthm commands, the hypothesis list is now an alternating
list of hypothesis name then hypothesis value, rather than a list
of (HYPNAME HYP_EXPR) pairs. This gets rid of some annoying parentheses.

The proof steps in a defthm or thm command are no longer enclosed in
parentheses as a single list; the are just added to the command argument
immediately after the list containing the conclusions.  This also
lightens the parentheses load.